refactored menu and linting
